Website/requirements.txt
Flask
langchain
langchain-huggingface
langchain-community
torch
sentence-transformers
ollama